[#608] Refine OtherToolsAndResources section into a component to load in PL and CII
