+++ frontend/src/components/utils/index.js
@@ -20,5 +20,6 @@ export { default as PasswordCriteria } from "./PasswordCriteria";
 export { default as ShowLabelButton } from "./ShowLabelButton";
 export { default as DataSecurityProvisionModal } from "./DataSecurityProvisionModal";
 export { default as ViewSummaryModal } from "./ViewSummaryModal";
-export { default as OtherToolResourceList } from "./OtherToolResourceList";
+export { default as OtherToolResourceList } from "./other-tools-and-resources/OtherToolResourceList";
 export { default as NewCpiForm } from "./NewCpiForm";
+export { default as OtherToolsAndResources } from "./other-tools-and-resources/OtherToolsAndResources";
