+++ frontend/src/pages/procurement-library/procurement-library.scss
@@ -88,7 +88,7 @@
 
       .jumbotron-cards {
         position: absolute;
-        bottom: -112px;
+        bottom: -124px;
         left: 0;
         width: 100%;
         padding: 94px 62px 0;
